Stable - Update 4.4.0

+++ modules/Settings/ModuleManager/models/Library.php
@@ -1,50 +1,46 @@
 <?php
 
 /**
- * Module Manager Library class
- * @package YetiForce.Model
- * @license licenses/License.html
+ * Module Manager Library class.
+ *
+ * @copyright YetiForce Sp. z o.o
+ * @license YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author Mariusz Krzaczkowski <m.krzaczkowski@yetiforce.com>
  */
 class Settings_ModuleManager_Library_Model
 {
-
 	/**
-	 * List of all installation libraries
-	 * @var array 
+	 * List of all installation libraries.
+	 *
+	 * @var array
 	 */
-	public static $dirs = [
-		'mPDF' => ['dir' => 'libraries/mPDF/', 'url' => 'https://github.com/YetiForceCompany/lib_mPDF', 'name' => 'lib_mPDF'],
-		'roundcube' => ['dir' => 'modules/OSSMail/roundcube/', 'url' => 'https://github.com/YetiForceCompany/lib_roundcube', 'name' => 'lib_roundcube'],
-		'PHPExcel' => ['dir' => 'libraries/PHPExcel/', 'url' => 'https://github.com/YetiForceCompany/lib_PHPExcel', 'name' => 'lib_PHPExcel'],
-		'AJAXChat' => ['dir' => 'libraries/AJAXChat/', 'url' => 'https://github.com/YetiForceCompany/lib_AJAXChat', 'name' => 'lib_AJAXChat']
+	public static $libraries = [
+		'mPDF' => ['dir' => 'vendor/mPDF/', 'url' => 'https://github.com/YetiForceCompany/lib_mPDF', 'name' => 'lib_mPDF'],
+		'roundcube' => ['dir' => 'public_html/modules/OSSMail/roundcube/', 'url' => 'https://github.com/YetiForceCompany/lib_roundcube', 'name' => 'lib_roundcube'],
 	];
 
 	/**
-	 * Path to save temporary files
-	 * @var string 
+	 * Path to save temporary files.
+	 *
+	 * @var string
 	 */
-	public static $tempDir = 'cache' . DIRECTORY_SEPARATOR . 'upload';
+	const TEMP_DIR = 'cache' . DIRECTORY_SEPARATOR . 'upload';
 
 	/**
-	 * Temporary table that contains checked library ststuses
-	 * @var array 
-	 */
-	public static $cache = [];
-
-	/**
-	 * Function to check library status
+	 * Function to check library status.
+	 *
 	 * @param string $name
-	 * @return boolean
+	 *
+	 * @return bool
 	 */
 	public static function checkLibrary($name)
 	{
-		if (isset(static::$cache[$name])) {
-			return static::$cache[$name];
+		if (App\Cache::has('LIBRARY', $name)) {
+			return App\Cache::get('LIBRARY', $name);
 		}
 		$status = true;
-		if (isset(static::$dirs[$name])) {
-			$lib = static::$dirs[$name];
+		if (static::$libraries[$name]) {
+			$lib = static::$libraries[$name];
 			if (file_exists($lib['dir'] . 'version.php')) {
 				$libVersion = require $lib['dir'] . 'version.php';
 				if (App\Version::check($libVersion['version'], $lib['name'])) {
@@ -52,17 +48,20 @@ class Settings_ModuleManager_Library_Model
 				}
 			}
 		}
-		static::$cache[$name] = $status;
+		App\Cache::save('LIBRARY', $name, $status, App\Cache::LONG);
+
 		return $status;
 	}
 
 	/**
-	 * Get a list of all libraries and their statuses
+	 * Get a list of all libraries and their statuses.
+	 *
 	 * @return array
 	 */
 	public static function &getAll()
 	{
-		foreach (static::$dirs as $name => &$lib) {
+		$libs = [];
+		foreach (static::$libraries as $name => $lib) {
 			$status = 0;
 			if (is_dir($lib['dir'])) {
 				$status = 2;
@@ -74,54 +73,52 @@ class Settings_ModuleManager_Library_Model
 				}
 			}
 			$lib['status'] = $status;
+			$libs[$name] = $lib;
 		}
-		return static::$dirs;
+		return $libs;
 	}
 
 	/**
-	 * Download all missing libraries
-	 * @throws \Exception\NoPermitted
+	 * Download all missing libraries.
 	 */
 	public static function downloadAll()
 	{
-		foreach (static::$dirs as $name => &$lib) {
+		foreach (static::$libraries as $name => &$lib) {
 			static::download($name);
 		}
 	}
 
 	/**
-	 * Function to download library
+	 * Function to download library.
+	 *
 	 * @param string $name
-	 * @return boolean
+	 *
 	 * @throws \Exception\NoPermitted
+	 *
+	 * @return bool
 	 */
 	public static function download($name)
 	{
-		if (!isset(static::$dirs[$name])) {
+		if (!static::$libraries[$name]) {
 			App\Log::warning('Library does not exist: ' . $name);
-			throw new \Exception\NoPermitted('LBL_PERMISSION_DENIED');
+			throw new \App\Exceptions\NoPermitted('LBL_PERMISSION_DENIED', 406);
 		}
-
-		$lib = static::$dirs[$name];
-		if (file_exists($lib['dir'] . 'version.php')) {
-			App\Log::info('Library has already been downloaded: ' . $name);
-			return false;
-		}
-		$path = static::$tempDir . DIRECTORY_SEPARATOR . $name . '.zip';
+		$lib = static::$libraries[$name];
+		$path = static::TEMP_DIR . DIRECTORY_SEPARATOR . $lib['name'] . '.zip';
 		$mode = AppConfig::developer('MISSING_LIBRARY_DEV_MODE') ? 'developer' : App\Version::get($lib['name']);
 		$compressedName = $lib['name'] . '-' . $mode;
-		if (!file_exists($path)) {
+		if (!file_exists($path) && \App\RequestUtil::isNetConnection()) {
 			stream_context_set_default([
 				'ssl' => [
-					'verify_peer' => false,
-					'verify_peer_name' => false,
+					'verify_peer' => true,
+					'verify_peer_name' => true,
 				],
 			]);
 			$url = $lib['url'] . "/archive/$mode.zip";
 			$headers = get_headers($url, 1);
 			if (isset($headers['Status']) && strpos($headers['Status'], '302') !== false) {
 				App\Log::trace('Started downloading library: ' . $name);
-				if ($file = file_get_contents($url, false, stream_context_create(['ssl' => ['verify_peer' => false, 'verify_peer_name' => false]]))) {
+				if ($file = file_get_contents($url, false, stream_context_create(['ssl' => ['verify_peer' => true, 'verify_peer_name' => true]]))) {
 					file_put_contents($path, $file);
 					App\Log::trace('Completed downloads library: ' . $name);
 				}
@@ -130,9 +127,9 @@ class Settings_ModuleManager_Library_Model
 			}
 		}
 		if (file_exists($path) && filesize($path) > 0) {
-			$unzip = new \vtlib\Unzip($path);
-			$unzip->unzipAllEx('.', [], [ $compressedName => $lib['dir']]);
-			$unzip->close();
+			\vtlib\Functions::recurseDelete($lib['dir']);
+			$zip = \App\Zip::openFile($path, ['checkFiles' => false]);
+			$zip->unzip([$compressedName => $lib['dir']]);
 			unlink($path);
 		} else {
 			App\Log::warning('No import file: ' . $name);
@@ -140,14 +137,12 @@ class Settings_ModuleManager_Library_Model
 	}
 
 	/**
-	 * Function to update library
+	 * Function to update library.
+	 *
 	 * @param string $name
-	 * @throws \Exception\NoPermitted
 	 */
 	public static function update($name)
 	{
-		$lib = static::$dirs[$name];
-		\vtlib\Functions::recurseDelete($lib['dir']);
 		static::download($name);
 	}
 }
